backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.9.0
httpx==0.27.0
python-multipart==0.0.6
openai==1.3.0
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pdf2image==1.16.3
easyocr==1.7.0